API

Modules

ChangesOfVariables.ChangesOfVariables

Types and constants

Functions and macros

ChangesOfVariables.with_logabsdet_jacobian

Documentation

ChangesOfVariables.ChangesOfVariables — Module

ChangesOfVariables

Lightweight package that defines functionality to calculate volume element changes for functions that perform a change of variables (like coordinate transformations).

ChangesOfVariables.with_logabsdet_jacobian — Function

with_logabsdet_jacobian(f, x)

Computes both the transformed value of x under the transformation f and the logarithm of the volume element.

For (y, ladj) = with_logabsdet_jacobian(f, x), the following must hold true:

y == f(x)

ladj is the log(abs(det(jacobian(f, x))))

with_logabsdet_jacobian comes with support for broadcasted/mapped functions (via Base.Fix1) and (Julia >=v1.6 only) ComposedFunction.

Example:

foo(x) = inv(exp(-x) + 1)

function ChangesOfVariables.with_logabsdet_jacobian(::typeof(foo), x)
    y = foo(x)
    ladj = -x + 2 * log(y)
    (y, ladj)
end

x = 4.2
y, ladj_y = with_logabsdet_jacobian(foo, x)

X = rand(10)
broadcasted_foo = Base.Fix1(broadcast, foo)
Y, ladj_Y = with_logabsdet_jacobian(broadcasted_foo, X)

# Requires Julia >= v1.6:
z, ladj_z = with_logabsdet_jacobian(log ∘ foo, x)
z == log(foo(x))
ladj_z == ladj_y + with_logabsdet_jacobian(log, y)[2]